app/requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
docker==4.0.2
docker-registry-client==0.5.2
ecdsa==0.13.2
Flask==1.1.1
Flask-Cors==3.0.8
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jws==0.1.3
MarkupSafe==1.1.1
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
urllib3==1.25.6
waitress==1.3.1
websocket-client==0.56.0
Werkzeug==0.16.0
